New Python Notebook and Temporary Data
Check Slack for specifics on what needs to be done. Dataset takes around 10 hours to make.

I called the file Megha_ just for lack of confusion purposes ~ I am bad at Github, so if you can find a way to replace the old notebook with this one that would be great!

Datafile must be in sample folder as notebook.